+++ bpdm-gate-api/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/api/model/RelationOutputDto.kt
@@ -26,5 +26,6 @@ data class RelationOutputDto(
     val relationType: SharableRelationType,
     val sourceBpnL: String,
     val targetBpnL: String,
+    val states: Collection<RelationStateDto>,
     val updatedAt: Instant
 )