+++ CHANGELOG.md
@@ -15,6 +15,7 @@ For changes to the BPDM Helm charts please consult the [changelog](charts/bpdm/C
 - BPDM Cleaning Dummy: Configuration and scheduling support for relation cleaning dummy service with unit test coverage [#1383](https://github.com/eclipse-tractusx/bpdm/issues/1383).
 - BPDM Pool: Add validation for too many identifiers when creating/updating business partners [1388](https://github.com/eclipse-tractusx/bpdm/issues/1388)
 - BPDM: Add IsOwnedBy relation to the golden record process for sharing majority-owned legal hierarchies [1399](https://github.com/eclipse-tractusx/bpdm/issues/1399)
+- BPDM: Add validity on business partner relation to the golden record process for legal entity relations [#1295](https://github.com/eclipse-tractusx/bpdm/issues/1295)
 
 ### Changed